1.9

+++ eladmin-system/src/main/java/me/zhengjie/modules/system/rest/UserController.java
@@ -4,9 +4,11 @@ import me.zhengjie.aop.log.Log;
 import me.zhengjie.config.DataScope;
 import me.zhengjie.domain.Picture;
 import me.zhengjie.domain.VerificationCode;
+import me.zhengjie.modules.system.domain.Role;
 import me.zhengjie.modules.system.domain.User;
 import me.zhengjie.exception.BadRequestException;
 import me.zhengjie.modules.system.service.DeptService;
+import me.zhengjie.modules.system.service.RoleService;
 import me.zhengjie.service.PictureService;
 import me.zhengjie.service.VerificationCodeService;
 import me.zhengjie.utils.*;
@@ -25,6 +27,7 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import java.util.*;
+import java.util.stream.Collectors;
 
 /**
  * @author jie
@@ -48,6 +51,9 @@ public class UserController {
 
     @Autowired
     private DeptService deptService;
+
+    @Autowired
+    private RoleService roleService;
 
     @Autowired
     private VerificationCodeService verificationCodeService;
@@ -96,6 +102,7 @@ public class UserController {
         if (resources.getId() != null) {
             throw new BadRequestException("A new "+ ENTITY_NAME +" cannot already have an ID");
         }
+        checkLevel(resources);
         return new ResponseEntity(userService.create(resources),HttpStatus.CREATED);
     }
 
@@ -103,6 +110,7 @@ public class UserController {
     @PutMapping(value = "/users")
     @PreAuthorize("hasAnyRole('ADMIN','USER_ALL','USER_EDIT')")
     public ResponseEntity update(@Validated(User.Update.class) @RequestBody User resources){
+        checkLevel(resources);
         userService.update(resources);
         return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
@@ -111,21 +119,27 @@ public class UserController {
     @DeleteMapping(value = "/users/{id}")
     @PreAuthorize("hasAnyRole('ADMIN','USER_ALL','USER_DELETE')")
     public ResponseEntity delete(@PathVariable Long id){
+        Integer currentLevel =  Collections.min(roleService.findByUsers_Id(SecurityUtils.getUserId()).stream().map(Role::getLevel).collect(Collectors.toList()));
+        Integer optLevel =  Collections.min(roleService.findByUsers_Id(id).stream().map(Role::getLevel).collect(Collectors.toList()));
+
+        if (currentLevel > optLevel) {
+            throw new BadRequestException("角色权限不足");
+        }
         userService.delete(id);
         return new ResponseEntity(HttpStatus.OK);
     }
 
     /**
      * 验证密码
-     * @param pass
+     * @param user
      * @return
      */
-    @GetMapping(value = "/users/validPass/{pass}")
-    public ResponseEntity validPass(@PathVariable String pass){
+    @PostMapping(value = "/users/validPass")
+    public ResponseEntity validPass(@RequestBody User user){
         UserDetails userDetails = SecurityUtils.getUserDetails();
         Map map = new HashMap();
         map.put("status",200);
-        if(!userDetails.getPassword().equals(EncryptUtils.encryptPassword(pass))){
+        if(!userDetails.getPassword().equals(EncryptUtils.encryptPassword(user.getPassword()))){
            map.put("status",400);
         }
         return new ResponseEntity(map,HttpStatus.OK);
@@ -133,16 +147,16 @@ public class UserController {
 
     /**
      * 修改密码
-     * @param pass
+     * @param user
      * @return
      */
-    @GetMapping(value = "/users/updatePass/{pass}")
-    public ResponseEntity updatePass(@PathVariable String pass){
+    @PostMapping(value = "/users/updatePass")
+    public ResponseEntity updatePass(@RequestBody User user){
         UserDetails userDetails = SecurityUtils.getUserDetails();
-        if(userDetails.getPassword().equals(EncryptUtils.encryptPassword(pass))){
+        if(userDetails.getPassword().equals(EncryptUtils.encryptPassword(user.getPassword()))){
             throw new BadRequestException("新密码不能与旧密码相同");
         }
-        userService.updatePass(userDetails.getUsername(),EncryptUtils.encryptPassword(pass));
+        userService.updatePass(userDetails.getUsername(),EncryptUtils.encryptPassword(user.getPassword()));
         return new ResponseEntity(HttpStatus.OK);
     }
 
@@ -175,5 +189,17 @@ public class UserController {
         verificationCodeService.validated(verificationCode);
         userService.updateEmail(userDetails.getUsername(),user.getEmail());
         return new ResponseEntity(HttpStatus.OK);
+    }
+
+    /**
+     * 如果当前用户的角色级别低于创建用户的角色级别，则抛出权限不足的错误
+     * @param resources
+     */
+    private void checkLevel(User resources) {
+        Integer currentLevel =  Collections.min(roleService.findByUsers_Id(SecurityUtils.getUserId()).stream().map(Role::getLevel).collect(Collectors.toList()));
+        Integer optLevel = roleService.findByRoles(resources.getRoles());
+        if (currentLevel > optLevel) {
+            throw new BadRequestException("角色权限不足");
+        }
     }
 }
